[refactor] 프로젝트 패키지를 개선하라
각 역할에 맞게 패키지를 만들어 클래스를 개선합니다
