+++ src/main/java/carracing/RacingGameApplication.java
@@ -1,0 +1,10 @@
+package carracing;
+
+import carracing.controller.RacingGameController;
+
+public class RacingGameApplication {
+	public static void main(String[] args) {
+		RacingGameController racingGameController = new RacingGameController();
+		racingGameController.race();
+	}
+}
