--- src/main/java/carracing/car/MovingStrategy.java
@@ -1,5 +1,0 @@
-package carracing.car;
-
-public interface MovingStrategy {
-	boolean isPossible();
-}
