--- src/main/java/carracing/car/RacingGame.java
@@ -1,13 +1,0 @@
-package carracing.car;
-
-public class RacingGame {
-	private final Cars cars;
-
-	public RacingGame(Cars cars) {
-		this.cars = cars;
-	}
-
-	public void race(MovingStrategy movingStrategy, int tryingTimes) {
-		cars.race(movingStrategy, tryingTimes);
-	}
-}
