--- src/main/java/carracing/car/WinnerStrategy.java
@@ -1,7 +1,0 @@
-package carracing.car;
-
-import java.util.List;
-
-public interface WinnerStrategy {
-	List<Car> winners(List<Car> cars);
-}
