--- src/main/java/carracing/controlstatement/GreaterThan.java
@@ -1,16 +1,0 @@
-package carracing.controlstatement;
-
-public class GreaterThan extends Boolean {
-	private final int left;
-	private final int right;
-
-	public GreaterThan(int left, int right) {
-		this.left = left;
-		this.right = right;
-	}
-
-	@Override
-	public boolean booleanValue() {
-		return left > right;
-	}
-}
