--- src/main/java/carracing/controlstatement/IF.java
@@ -1,14 +1,0 @@
-package carracing.controlstatement;
-
-public class IF extends Boolean {
-	private final Boolean aBoolean;
-
-	public IF(Boolean aBoolean) {
-		this.aBoolean = aBoolean;
-	}
-
-	@Override
-	public boolean booleanValue() {
-		return aBoolean.booleanValue();
-	}
-}
