+++ src/main/java/carracing/domain/game/MovingStrategy.java
@@ -1,0 +1,5 @@
+package carracing.domain.game;
+
+public interface MovingStrategy {
+	boolean isPossible();
+}
