+++ src/main/java/carracing/domain/game/WinnerStrategy.java
@@ -1,0 +1,9 @@
+package carracing.domain.game;
+
+import carracing.domain.car.Car;
+
+import java.util.List;
+
+public interface WinnerStrategy {
+	List<Car> winners(List<Car> cars);
+}
