+++ src/main/java/carracing/util/controlstatement/IF.java
@@ -1,0 +1,14 @@
+package carracing.util.controlstatement;
+
+public class IF extends Boolean {
+	private final Boolean aBoolean;
+
+	public IF(Boolean aBoolean) {
+		this.aBoolean = aBoolean;
+	}
+
+	@Override
+	public boolean booleanValue() {
+		return aBoolean.booleanValue();
+	}
+}
