+++ src/main/java/carracing/util/controlstatement/OR.java
@@ -1,0 +1,16 @@
+package carracing.util.controlstatement;
+
+public class OR extends Boolean {
+	private final Boolean aBoolean;
+	private final Boolean bBoolean;
+
+	public OR(Boolean aBoolean, Boolean bBoolean) {
+		this.aBoolean = aBoolean;
+		this.bBoolean = bBoolean;
+	}
+
+	@Override
+	public boolean booleanValue() {
+		return this.aBoolean.booleanValue() || this.bBoolean.booleanValue();
+	}
+}
