+++ src/main/java/carracing/view/InputView.java
@@ -1,13 +1,16 @@
 package carracing.view;
 
-import carracing.car.Car;
+import carracing.domain.car.Car;
+import carracing.domain.car.Cars;
 
 import java.util.ArrayList;
 import java.util.List;
 import java.util.Scanner;
 
 public class InputView {
-	public static List<Car> inputCars(Scanner scanner) {
+	private final static Scanner scanner = new Scanner(System.in);
+
+	public static Cars inputCars() {
 		System.out.println("경주할 자동차 이름을 입력하세요(이름은 쉼표(,)를 기준으로 구분).");
 		String[] inputCarNames = scanner.nextLine().split(",");
 
@@ -16,10 +19,10 @@ public class InputView {
 			cars.add(new Car(inputCarName));
 		}
 
-		return cars;
+		return new Cars(cars);
 	}
 
-	public static int inputTryingNumber(Scanner scanner) {
+	public static int inputTryingNumber() {
 		System.out.println("시도할 회수는 몇 회 인가요?");
 		int tryingNumber = scanner.nextInt();
 		return tryingNumber;
